requirements.txt
altgraph==0.17.2
cupy-cuda115==9.6.0
Cython==0.29.32
fastrlock==0.8
importlib-metadata==4.8.3
llvmlite==0.36.0
numba==0.53.1
numpy==1.19.5
pkg_resources==0.0.0
scipy==1.5.4
typing_extensions==4.1.1
zipp==3.6.0
